Projection for the pre fixed bonds.
This projection aim to show the current value of a bond prefixed.

Fix the projection precision. It was fixed changing to "double" from "float".

Partitioning of the code that make the financial projection enabling it reuse.

+++ TaxRequester/BaseTaxData.cs
@@ -4,14 +4,14 @@ namespace TaxRequester
 {
     public abstract class BaseTaxData
     {
-        public abstract float Tax { get; set; }
-        public virtual float GetTaxDaily(float inputTax = 0f)
+        public abstract double Tax { get; set; }
+        public virtual double GetTaxDaily(double inputTax = 0f)
         {
             if (inputTax == 0f)
             {
                 inputTax = Tax;
             }
-            return (float)Math.Pow(1 + inputTax / 100, 1f / 252) - 1;
+            return Math.Pow(1 + inputTax / 100, 1f / 252) - 1;
         }
     }
 }
